Complete 1 and 2

+++ src/sample/models/Category.java
@@ -6,19 +6,19 @@ import java.sql.SQLException;
 
 public class Category extends Categories {
 
-    private int id;
+    private long id;
     private String name;
 
-    public Category(int id, String name) throws SQLException, ClassNotFoundException {
+    public Category(long id, String name) throws SQLException, ClassNotFoundException {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
