+++ src/sample/models/Provider.java
@@ -6,19 +6,19 @@ import java.sql.SQLException;
 
 public class Provider extends Providers {
 
-    private int id;
+    private long id;
     private String name;
 
-    public Provider(int id, String name) throws SQLException, ClassNotFoundException {
+    public Provider(long id, String name) throws SQLException, ClassNotFoundException {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
